requirements.txt
ultralytics==8.1.0
torch==2.2.0+cpu
torchvision==0.17.0+cpu
torchaudio==2.2.0+cpu
opencv-python-headless==4.8.1.78
numpy==1.24.4
pyserial==3.5

